% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{astats_major}
\alias{astats_major}
\title{MCAT and GPA for all 2021-2022 medical school applicants by major.}
\format{
A data frame of 8 rows and 18 variables
\describe{
\item{major}{(character) Undergraduate major.}
\item{CPBS_mean}{(numeric) Mean score for the Chemical and Physical Foundations sub-section of the MCAT exam.}
\item{CPBS_SD}{(numeric) Standard deviation score for the Chemical and Physical Foundations sub-section of the MCAT exam.}
\item{CARS_mean}{(numeric) Mean score for the Critical Analysis and Reasoning sub-section of the MCAT exam.}
\item{CARS_SD}{(numeric) Standard deviation score for the Critical Analysis and Reasoning sub-section of the MCAT exam.}
\item{BBLS_mean}{(numeric) Mean score for the Biological and Biochemical Foundations sub-section of the MCAT exam.}
\item{BBLS_SD}{(numeric) Standard deviation score for the Biological and Biochemical Foundations sub-section of the MCAT exam.}
\item{PSBB_mean}{(numeric) Mean score for the Psychological and Sociological Foundations sub-section of the MCAT exam.}
\item{PSBB_SD}{(numeric) Standard deviation score for the Psychological and Sociological Foundations sub-section of the MCAT exam.}
\item{MCAT_mean}{(numeric) Mean total MCAT exam score.}
\item{MCAT_SD}{(numeric) Standard deviation of MCAT exam score.}
\item{sGPA_mean}{(numeric) Mean science grade point average (includes mathematics, physics, chemistry, biochemistry, and biology classes.}
\item{sGPA_SD}{(numeric) Standard deviation of science grade point average}
\item{aoGPA_mean}{(numeric) Mean grade point average for 'all-other' (i.e. non-science) classes.}
\item{aoGPA_SD}{(numeric) Standard deviation of grade point average for all-other classes.}
\item{cGPA_mean}{(numeric) Mean cumulative grade point average.}
\item{cGPA_SD}{(numeric) Standard deviation of cumulative grade point average.}
\item{num_apps}{(numeric) Number of applicants.}
}
}
\source{
2021 FACTS: Applicants and Matriculants Data, Association of American Medical Colleges
\url{https://www.aamc.org/data-reports/students-residents/interactive-data/2021-facts-applicants-and-matriculants-data}.
}
\usage{
astats_major
}
\description{
Mean total MCAT exam/sub-section scores and combined/science/non-science GPAs for all 2021-2022 applicants by undergraduate major
}
\keyword{datasets}
