% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mstats_race}
\alias{mstats_race}
\title{MCAT and GPA for all 2021-2022 medical school matriculants by race.}
\format{
A data frame of 99 rows and 5 variables
\describe{
\item{category}{(character) Category of measurement, e.g. total MCAT exam score or non-science GPA.}
\item{race}{(character) Race and ethnicity.}
\item{mean}{(numeric) Average value of category's statistic.}
\item{SD}{(numeric) Standard deviation of category's statistic.}
\item{num_apps}{(numeric) Number of applicable observations.}
}
}
\source{
2021 FACTS: Applicants and Matriculants Data, Association of American Medical Colleges
\url{https://www.aamc.org/data-reports/students-residents/interactive-data/2021-facts-applicants-and-matriculants-data}.
}
\usage{
mstats_race
}
\description{
Mean total MCAT exam/sub-section scores and combined/science/non-science GPAs for all 2021-2022 matriculants by race and ethnicity
}
\keyword{datasets}
